% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ada.R
\name{ada}
\alias{ada}
\title{Ancestral Diversity Analysis for assemblage data}
\usage{
ada(
  x,
  phy,
  node.names = FALSE,
  type = c("discrete", "continuous"),
  marginal = FALSE,
  sp.bin = 15,
  lik.threshold = FALSE,
  threshold = 0.2,
  compute.node.by.sites = FALSE,
  compute.fields = F,
  allowSplit = T,
  credMass = 0.95,
  compute.metrics = FALSE
)
}
\arguments{
\item{x}{}

\item{phy}{}

\item{node.names}{}

\item{type}{}

\item{marginal}{}

\item{sp.bin}{}

\item{lik.threshold}{}

\item{threshold}{}

\item{compute.node.by.sites}{}

\item{compute.fields}{}

\item{allowSplit}{}

\item{credMass}{}

\item{compute.metrics}{}
}
\description{
Ancestral Diversity Analysis for assemblage data
}
