% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/funciton_ada_core.R
\name{ada_core}
\alias{ada_core}
\title{Computation of core ADA components}
\usage{
ada_core(
  x,
  phy,
  type = c("discrete", "continuous"),
  lik.threshold = FALSE,
  threshold = 0.2,
  compute.node.by.sites = TRUE
)
}
\arguments{
\item{x}{community matrix. Sites are rows and species are columns}

\item{phy}{phylogenetic tree as newick object}

\item{type}{character indicating the type of reconstruction to be used. Default is continuous}

\item{lik.threshold}{Logical, default is FALSE}

\item{threshold}{scalar indicating the threshold to be used when considering the presence of a lineage in a site}

\item{compute.node.by.sites}{Logical, TRUE (default) computes a matrix of node occurrence by site}
}
\value{
a list with three elements. reconstruction is the result of ancestral area reconstruction; phylogeny is the matrix containing
the occurrence of nodes in sites and joint.phylo.obs is the joint occurrence of nodes and species in phylonetic tree
}
\description{
Computation of core ADA components
}
