% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliation_evoreg.R
\name{affiliation_evoreg}
\alias{affiliation_evoreg}
\title{Affiliation values for assemblages according to phylogenetic turnover}
\usage{
affiliation_evoreg(phylo.comp.dist, groups)
}
\arguments{
\item{phylo.comp.dist}{A distance matrix indicating the phylogenetic (or taxonomic/functional) distance composition
among assemblages}

\item{groups}{A character vector indicating the group of each assemblage}
}
\value{
A list with two matrix, one containing affiliation values and the group in which each cell
is classified and the other containing cell coordinates
}
\description{
Affiliation values for assemblages according to phylogenetic turnover
}
\examples{


}
