% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliation_evoreg.R
\name{afilliation.evoreg}
\alias{afilliation.evoreg}
\title{Affiliation values for assemblages according to phylogenetic turnover}
\usage{
afilliation.evoreg(evo.classification, distance = NULL, method = "euclidean")
}
\arguments{
\item{evo.classification}{An object returned from \code{\link{evoregions}} function or phyloregion}

\item{distance}{A distance matrix, only applies if the object in evo.classification is from class phyloregion}

\item{method}{Character indicating the method used to compute the distance in fuzzy matrix}
}
\value{
A list with two matrix, one containing affiliation values and the group in which each cell
is classified and the other containing cell coordinates
}
\description{
Affiliation values for assemblages according to phylogenetic turnover
}
