% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_age_arrival.R
\name{age_arrival}
\alias{age_arrival}
\title{Computes the arrival ages of each species in the assemblages}
\usage{
age_arrival(W, tree, ancestral.area, biogeo, age.no.ancestor = "recent")
}
\arguments{
\item{W}{Occurrence matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic tree in newick format}

\item{ancestral.area}{One column data frame, nodes in row and one column containing the occurrence of the nodes}

\item{biogeo}{One column data frame, assemblages in rows and one column containing the region in which the assemblage is located}

\item{age.no.ancestor}{a character string "recent" (default) or "half.edge" indicating how to deal
with cases in which the most recent aancestor of a tip node are not on the same region. "recent"
attributes a small age (10e-5), while "half.edge" attributes the age as half the length of the branch
linking the ancestor to the tip.}
}
\value{
A list of length two. \item{age_arrival}{A matrix with the arrival age of each species in each assemblage}
\item{mean_age_arrival}{mean age values for each assemblage}
}
\description{
Computes the arrival ages of each species in the assemblages
}
\author{
Gabriel Nakamura \href{mailto:gabriel.nakamura.souza@gmail.com}{gabriel.nakamura.souza@gmail.com} and Arthur Rodrigues
}
