% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{akodon.pa.tree}
\alias{akodon.pa.tree}
\title{Occurrence data for Akodon species}
\format{
data frame with 732 rows and 30 columns organized in the same sequence as the species in the
Akodon phylogenetic tree
}
\usage{
akodon.pa.tree
}
\description{
This is a filtered and ordered data frame containing the occurrence of 30 species from Akodon genus in 732 assemblages
}
\keyword{datasets}
