% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{akodon.sites}
\alias{akodon.sites}
\title{Occurrence data for Akodon species}
\format{
data frame containing 732 rows and 40 columns
}
\usage{
akodon.sites
}
\description{
This is a occurrence data for species from Akodon genus
}
\keyword{datasets}
