% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{akodon_newick}
\alias{akodon_newick}
\title{Phylogenetic tree in of Akodon genus}
\format{
newick phylogenetic tree
}
\usage{
akodon_newick
}
\description{
This is a phylogenetic tree containing 30 species from genus Akodon
}
\keyword{datasets}
