% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{averaged_rates}
\alias{averaged_rates}
\title{Averaged rates of life mode for Sigmodontinae assemblages}
\format{
An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 1770 rows and 3 columns.
}
\usage{
averaged_rates
}
\description{
Data frame with three columns with averaged values of tip based metrics of trait evolution
}
\keyword{datasets}
