% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dispersal_from.R
\name{calc_dispersal_from}
\alias{calc_dispersal_from}
\title{Proportional contribution of an ancestral area to the species in other region}
\usage{
calc_dispersal_from(W, tree, ancestral.area, biogeo)
}
\arguments{
\item{W}{Occurrence matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic tree in newick format}

\item{ancestral.area}{One column data frame with nodes in rows and one column indicating the occurrence (biome/ecoregion) area of nodes}

\item{biogeo}{One column data frame with assemblages in rows and their respective biome/ecoregion}
}
\value{
A data frame with assemblages in the rows and regions in the columns. The values indicates
the percentage of contribution of each region to each assemblage. NA represents no contribution
}
\description{
For the species present in a site, the function calculates the proportional
contribution of ancestral areas to dispersal of lineage in the site's region.
}
\examples{
\dontrun{
data(akodon.sites) # occurrence matrix
data(akodon.pa.tree) # phylogenetic tree
data(regions) # biogeographic region
data(resDEC) # output from ancestral area reconstruction
node.area <- get_node_range_BioGeoBEARS(resDEC,
                                        phyllip.file = here("inst", "extdata", "geo_area_akodon.data")
                                        ,akodon.newick,max.range.size = 4) # data frame indicating the ancestral area
calc_dispersal_from(W=akodon.sites,
                    tree=akodon.pa.tree,
                    ancestral.area=node.area,biogeo=regions) # historical dispersal analysis
}

}
\seealso{
\code{\link{get_node_range_BioGeoBEARS}}
}
\author{
Arthur V Rodrigues \href{mailto:rodrigues.arthur.v@gmail.com}{rodrigues.arthur.v@gmail.com}
}
