% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ed.R
\name{calc_ed}
\alias{calc_ed}
\title{Calculate evolutionary distinctiveness (ED) based on a biogeographical model}
\usage{
calc_ed(
  tree,
  ancestral.area = NULL,
  current.area = NULL,
  type = c("equal.splits", "fair.proportion")
)
}
\arguments{
\item{tree}{Phylogenetic tree of class \code{'phylo'}.}

\item{ancestral.area}{A one-column data frame indicating the area of occurrence
of each node (rows). Row names must correspond to node labels in the tree.}

\item{current.area}{A character string indicating the focal area. All tips
are assumed to be present in this area when computing in situ ED.}

\item{type}{Character indicating the type of ED metric to use. One of
\code{"equal.splits"} (default) or \code{"fair.proportion"}.}
}
\value{
A named numeric vector with ED values for all tips in the phylogeny.
}
\description{
For each tip, the function calculates evolutionary distinctiveness (ED), which
can be the total ED or in situ ED. See details. By default, it calculates total ED.
}
\details{
\itemize{
\item \strong{Total ED} (no biogeographical restriction):
If neither \code{ancestral.area} nor \code{current.area} are provided, the function
calculates ED as originally proposed by Redding & Mooers (2006, 2007).
\item \strong{In situ ED} (biogeographical restriction):
If both \code{ancestral.area} and \code{current.area} are provided, ED is calculated
only along branches where the ancestral area matches the provided \code{current.area}.
This represents evolutionary distinctiveness accumulated \emph{in situ} within
the specified biogeographical region.
}

You must provide \strong{either both} \code{ancestral.area} and \code{current.area} (for
in situ ED), \strong{or neither} (for total ED). Providing only one of them will
result in an error.
}
\examples{

# example for calc_ed
# generate simlutated data ----
set.seed(4523)
tree_sim <- ape::rcoal(5)

# Create node area table (only bifurcating nodes)
node_area <- data.frame(
    area = c("A", "A", "BC", "AD"),
    row.names = paste0("N", 6:9)
    )
    
# create insertion of nodes to simulate for change in area in the branch 
# Use ggtree to identify insertion points 

gdata <- ggtree::ggtree(tree_sim)$data
ins_data <- gdata \%>\% dplyr::filter(node \%in\% c(3, 9, 8))
ins_data <- tibble::add_row(ins_data, ins_data[2, ]) \%>\% arrange(node)

# Define inserted nodes with areas
inserts <- tibble::tibble(
  parent = ins_data$parent,
  child = ins_data$node,
  event_time = c(0.2, 0.5, 0.9, .3),
  node_area = c("AC", "AB", "BC", "D")
  )
  
# Insert nodes 
result <- insert_nodes(tree_sim, inserts, node_area = node_area)
tree_out <- result$phylo
node_area_out <- result$node_area


# Compute Evolutionary Distinctiveness --------

# ED total (same results as picante::evol.distinct())
ed_total <- calc_ed(  
  tree_out, 
  node_area = node_area_out, 
  type = "equal.splits")
  
# ED partial. Account for in situ distinctiveness only

ed_partial_A <- calc_ed(
  tree_out,
  node_area = node_area_out, 
  current_area = "A", 
  type = "equal.splits")
  
ed_partial_D <- calc_ed(
  tree_out,
  node_area = node_area_out, 
  current_area = "D", 
  type = "equal.splits")
  
data.frame(
  species = names(ed_total), 
  ed_total,
  ed_partial_A,
  ed_partial_D
  )
  
# tree visualization
tree_viz <- tree_out
tree_viz$node.label <- node_area_out$area

plot(tree_viz, show.node.label = T)
nodelabels(adj = 1.2)



}
\references{
Redding, D. W., & Mooers, A. Ø. (2006). Incorporating evolutionary measures into
conservation prioritization. \emph{Conservation Biology}, 20(6), 1670–1678.
https://doi.org/10.1111/j.1523-1739.2006.00555.x

Redding, D. W., & Mooers, A. Ø. (2007). The shape of phylogenetic trees and the
context for conservation: a review of macroevolutionary macroecology.
\emph{Philosophical Transactions of the Royal Society B: Biological Sciences}, 362(1478), 849–860.
https://doi.org/10.1098/rstb.2006.1977
}
