% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_insitu_PD_PE_metrics.R
\name{calc_insitu_metrics}
\alias{calc_insitu_metrics}
\title{Diversification-based PD and PE}
\usage{
calc_insitu_metrics(W, tree, ancestral.area, biogeo, PD = TRUE, PE = TRUE)
}
\arguments{
\item{W}{Occurrence matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic hipothesis in newick format}

\item{ancestral.area}{One column data frame with nodes in rows and one column indicating the occurrence area of nodes}

\item{biogeo}{One columns data frame with rows indicating assemblages and one column indicating the biome/ecoregion of each assemblage}

\item{PD}{Logical, if TRUE (default) PD~in-situ~ will be computed}

\item{PE}{Logical, if TRUE (default) PE~in-situ~ will be computed}
}
\value{
A data frame containing the values of original PD and PE and also their in-situ diversification
based version
}
\description{
Diversification-based PD and PE
}
\details{
This function computes two modified versions of Phylogenetic Diversity (PD) and Phylogenetic
endemism by considering in the calculation the in-situ diversification. the PD~in-situ~ and PE~in-situ~ are calculated
by coupling in the calculation an ancetral area reconstruction that divide the phylogenetic tree in two parts,
one that correspond to dispersion events and another that correspond to in-situ diversification events.
We use the in-situ diversification part to calculate both Db-PD and Db-PE
}
\examples{
W_toy<- matrix(c(0, 1, 1, 0, 1, 1, 0, 1, 0, 1, 0, 0, 1, 0, 0),
nrow= 3,
ncol= 5,
dimnames=list(c("Comm 1", "Comm 2", "Comm 3"),
c(paste("s", 1:5, sep=""))))
 data("toy_treeEx")
 biogeo_toy <- data.frame(Ecoregion= c("A", "B", "C"))
 ancestral_area_toy <- data.frame(state= c("ABC", "B", "C", "ABC"))
 assemblage_phylo_metrics <- calc_insitu_metrics(W_toy,
      toy_treeEx,
      ancestral_area_toy, 
      biogeo_toy)
}
\author{
Gabriel Nakamura \href{mailto:gabriel.nakamura.souza@gmail.com}{gabriel.nakamura.souza@gmail.com}
}
