% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_calc_PDada.R
\name{calc_PD_ada}
\alias{calc_PD_ada}
\title{Calculation of PD using ada object for one community}
\usage{
calc_PD_ada(ada.obj, threshold = 0.6, comm.name)
}
\arguments{
\item{ada.obj}{}

\item{threshold}{}

\item{comm.name}{}
}
\description{
Calculation of PD using ada object for one community
}
