% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{comm_sbears}
\alias{comm_sbears}
\title{Community composition of Sigmodontinae species}
\format{
A data frame object with 46 rows and 260 columns, each column
correspond to a species name and the rows to sites
}
\source{
\url{https://zenodo.org/records/16933539}
}
\usage{
comm_sbears
}
\description{
Community composition matrix in dense format with presence/absence of
260 species in 46 sites
}
\keyword{datasets}
