% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_comp_ada_nodes_sites.R
\name{comp_ada_nodes_sites}
\alias{comp_ada_nodes_sites}
\title{Axiliary function to compute node composition by sites}
\usage{
comp_ada_nodes_sites(phy, comm, long = FALSE)
}
\arguments{
\item{phy}{A phylo object}

\item{comm}{A community matrix, rows are communities and columns are species}

\item{long}{Logical, if TRUE dense format of the output matrix is converted
to long format}
}
\value{
A matrix (or data frame in the long format) with community composition
by nodes
}
\description{
Axiliary function to compute node composition by sites
}
