% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_comp-node-age.R
\name{comp_age}
\alias{comp_age}
\title{Compute node ages}
\usage{
comp_age(phy)
}
\arguments{
\item{phy}{a phylogenetic tree with nodes named}
}
\value{
a data frame with two columns, one with species and node names, another with node ages
species will show age 0
}
\description{
Compute node ages
}
