% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_kernel_gravel_density_dist_decay.R
\name{comp_kernel_gravel}
\alias{comp_kernel_gravel}
\title{Occupancy probability of each node in each site considering a kernel density funcion}
\usage{
comp_kernel_gravel(x, phy, coords, w_slope, min_disp_prob)
}
\arguments{
\item{x}{Node by communities matrix}

\item{phy}{phylo object}

\item{coords}{data frame with spatial coordinate of communities}
}
\description{
Occupancy probability of each node in each site considering a kernel density funcion
}
