% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model_based_diversificationJetz.R
\name{db_diversification}
\alias{db_diversification}
\title{Tip-based and model-based diversification metrics}
\usage{
db_diversification(
  W,
  tree,
  ancestral.area,
  biogeo,
  diversification = "jetz",
  type = "equal.splits"
)
}
\arguments{
\item{W}{Assemblage occurrence matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic tree in newick format}

\item{ancestral.area}{One column data frame indicating the Ecoregion of occurrence of each node (rows)}

\item{biogeo}{Data frame with one column indicating the Ecoregion of each assemblage}

\item{diversification}{Character indicating the tip-based diversification metric to be calculated. Default is "jetz"}

\item{type}{Character indicating the type of calculation to be used in the calculation of tip-based metric, "equal-splits" is the default}
}
\value{
A list with the following objects:
\itemize{
\item{Jetz_per_spp}{Jetz tip-based metric of diversification for each species}
\item{Jetz_species_site}{Matrix containing
the values of Jetz tip-based diversification metric for each assemblage}
\item{Jetz_harmonic_mean_site}{Harmonic mean of Jetz tip-based
diversification metric}
\item{model_based_Jetz_species_sites}{Matrix with model-based Jetz metric for each species
in each assemblage}
\item{model_based_Jetz_harmonic_mean_site}{Harmonic mean of model-based Jetz diversification}
}
}
\description{
Tip-based and model-based diversification metrics
}
\details{
model-based measure of diversification is calculated as the proportion of total tip-based diversification
metric corresponded only to the branches of a given species that diversified in the current area of occurrence of that species
\deqn{Jetz_total \times (ED_div/ED_total)}
}
\author{
Gabriel Nakamura \href{mailto:gabriel.nakamura.souza@gmail.com}{gabriel.nakamura.souza@gmail.com} and Arthur V Rodrigues
}
