% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_Db-PD_PE_metrics.R
\name{div_based_metrics}
\alias{div_based_metrics}
\title{Diversification-based PD and PE}
\usage{
div_based_metrics(W, tree, ancestral.area, biogeo, PD = TRUE, PE = TRUE)
}
\arguments{
\item{W}{Occurrence matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic hipothesis in newick format}

\item{ancestral.area}{One column data frame with nodes in rows and one column indicating the occurrence area of nodes}

\item{biogeo}{One columns data frame with rows indicating assemblages and one column indicating the biome/ecoregion of each assemblage}

\item{PD}{Logical, if TRUE (default) Db-PD will be computed}

\item{PE}{Logical, if TRUE (default) Db-PE will be computed}
}
\value{

}
\description{
Diversification-based PD and PE
}
\author{
Gabriel Nakamura \href{mailto:gabriel.nakamura.souza@gmail.com}{gabriel.nakamura.souza@gmail.com}
}
