% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivB_metrics.R
\name{DivB_metrics}
\alias{DivB_metrics}
\title{Diversification Based (DivB) Metrics for Community Phylogenetics}
\usage{
DivB_metrics(
  W,
  tree,
  ancestral.area,
  biogeo,
  diversification = c("jetz", "freck"),
  PD = TRUE,
  PE = TRUE,
  age.arrival = TRUE,
  age.no.ancestor = NA,
  dispersal.from = TRUE,
  ED.type = "equal.splits"
)
}
\arguments{
\item{W}{community matrix (site x species)}

\item{tree}{phylogenetic tree}

\item{ancestral.area}{biogeographic area for ancestral. Must be a single
column data.frame in which rows are nodes in the tree. Each area must be a
single letter that could be combined (eg. "A", "AB", "B" ).}

\item{biogeo}{sinlge collum data.frame with biogeographic area for each site.}

\item{diversification}{character. Diversification index sensu Jetz et al. or
sensu Frecklerton et al. Defaut = c("jetz", "freck").}

\item{PD}{logical. calculate phylogenetic diversity and its diversification based version (PDlocal). Defaut = TRUE.}

\item{PE}{logical. calculate phylogenetic endemism and its diversification based version (Db-PE). Defaut = TRUE.}

\item{age.arrival}{logical. Calculate mean age of arrival of species in the
site's biogeographic area. Defaut = TRUE.}

\item{age.no.ancestor}{how to deal with age arrival for species in which the
late ancestor wasn't in the site's biogeographic area. Options: NA,
numeric(), or 'half.edge'. Defalt = NA.  See Details.}

\item{dispersal.from}{logical. Calculate frequecy of dipersal from each
biogeografic area of ancestors for each site}
}
\description{
Compute in each site DivB metrics for diversification (Jetz and Frecklerton),
diversity metrics (Phylogenenic diversity and phylogenetic endemism),
age of arrival in the biogeographic area, and frequency of dispersal
from other areas
}
