% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoregions.R
\name{evoregions}
\alias{evoregions}
\title{Defining groups based on phylogenetic turnover}
\usage{
evoregions(
  comm,
  coords,
  phy,
  method.dist = "bray",
  tresh.dist = 0.05,
  method.clust = "kmeans",
  stat.clust = "BIC",
  n.iter.clust = 1e+07,
  criterion.clust = "diffNgroup",
  max.n.clust = 10
)
}
\arguments{
\item{comm}{Species occurrence matrix. Assemblages are rows and species are columns}

\item{coords}{Numerical matrix with coordinates for assemblages in comm}

\item{phy}{Newick object with phylogenetic tree containing the species in comm}

\item{method.dist}{Character. The method to be used to compute phyogenetic distances among assemblages.
Default is "bray"}

\item{tresh.dist}{A scalar informing the threshold value to select eigenvectors based on the amount of variation of each eigenvector.
Default is 0.05 (eigenvector with at least 5\% of variation)}

\item{method.clust}{Character indicating the grouping algorithm to be used in cluster analysis.
Defalt is "kmeans"}

\item{stat.clust}{Character to be used in \code{\link{find.clusters}} indicating the statistic to be computed for each model. Can be
"BIC" (default), "AIC" or "WSS".}

\item{n.iter.clust}{Integer to be used in \code{\link{find.clusters}}
function of adegenet package to indicate the number of iterations to be used in each run of K-means algorithm}

\item{criterion.clust}{Character}

\item{max.n.clust}{Integer value to be used in \code{\link{find.clusters}}.
Indicates the maximum number of clusters to be tried.}
}
\value{
A list of length four containing
1 - Kstat: Numeric vector with the values of the summary statistics for different K values
2 - stat: Numeric value representing the summary statistic for the retained model
3 - grp: A list with three components, the K statistic for each group, the group that each
assemblage was classified
4 - Size: The number of assemblages in each group
}
\description{
Defining groups based on phylogenetic turnover
}
