% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_find_maxn_clust.R
\name{find.max.nclust}
\alias{find.max.nclust}
\title{Estimate the maximum number of groups in DAPC analysis}
\usage{
find.max.nclust(
  x,
  threshold,
  max.nclust,
  nperm = 100,
  method = "kmeans",
  stat = "BIC",
  criterion = "diffNgroup",
  subset = 100,
  confidence.level = c(0.7, 0.8, 0.9, 0.95, 0.99)
)
}
\arguments{
\item{x}{A data.frame or matrix object containing eigenvectors by sites data.}

\item{threshold}{Scalar. The number of eigenvectors used to perform classification.}

\item{max.nclust}{Value of set of values defining the maximum number of groups to be evaluated.}

\item{nperm}{Scalar. Number of times classification will be performed.}

\item{method}{c("kmeans","ward"). See \code{\link{find.clusters}} of adegenet package.}

\item{stat}{c("BIC", "AIC", or "WSS"). See \code{\link{find.clusters}} of adegenet package.}

\item{criterion}{c("diffNgroup", "min","goesup", "smoothNgoesup", or "goodfit"). See \code{\link{find.clusters}} of adegenet package.}

\item{subset}{number of cells used in the analysis. It is particularly important whenever the total number of cells is large (> 1000).}

\item{confidence.level}{threshold confidence level used to estimate congruence in the classification pattern.}
}
\value{
Matrix containing congruence values ranging between 0-1 for each max.nclust value (see Arguments) and confidence level.
}
\description{
Estimate the maximum number of groups in DAPC analysis
}
\examples{

}
