% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getNodeRange_biogeobears.R
\name{get.node.range_BioGeoBEARS}
\alias{get.node.range_BioGeoBEARS}
\title{Get node ranges from BioGeoBEARS biome reconstruction model}
\usage{
get.node.range_BioGeoBEARS(
  BioGeoBEARS.data,
  phyllip.file,
  tree,
  max.range.size
)
}
\arguments{
\item{BioGeoBEARS.data}{An object containing the result of BioGeoBEARS reconstruction}

\item{phyllip.file}{Phylip file, the same used in BioGeoBEARS package}

\item{tree}{A phylogenetic tree}

\item{max.range.size}{Scalar indicating the maximum number of biomes in which each node can occupy}
}
\value{
Data frame with two columns, one indicating the node and other the ancestral biome range
}
\description{
Get node ranges from BioGeoBEARS biome reconstruction model
}
\examples{

}
