% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_bsm_node_area.R
\name{get_bsm_node_area}
\alias{get_bsm_node_area}
\title{Get the node area states from the biogeographical stochastic mapping's result}
\usage{
get_bsm_node_area(
  bsm,
  BioGeoBEARS.data,
  phyllip.file,
  tree.path,
  max.range.size
)
}
\arguments{
\item{bsm}{results from the function \code{\link[=calc_bsm]{calc_bsm()}}}

\item{BioGeoBEARS.data}{a BioGeoBEARS result model object}

\item{phyllip.file}{path to the phyllip file used in the original model}

\item{tree.path}{path for the phylogenetic tree used in the original model}

\item{max.range.size}{maximum range size used in the biogeographical reconstruction}
}
\value{
a list with same length as the number of stochastic mapping. Each
list element is a single column data frame with node area states
}
\description{
Get the node area states from the biogeographical stochastic mapping's result
}
