% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_insert_nodes_df.R
\name{get_insert_df}
\alias{get_insert_df}
\title{Get data frame of transition states from Biogeographical Stochastic Mapping}
\usage{
get_insert_df(bsm, phyllip.file, max.range.size, include_null_range = TRUE)
}
\arguments{
\item{bsm}{results from the function \code{\link[=calc_bsm]{calc_bsm()}}}

\item{phyllip.file}{path to the phyllip file used in the original model}

\item{max.range.size}{maximum range size used in the biogeographical reconstruction}

\item{include_null_range}{default TRUE.}
}
\value{
a list with same length as the number of stochastic mapping. Each
list element is a data frame with columns:
\itemize{
\item \code{child} = child node number,
\item \code{parent} = ancestor node number,
\item \code{event_time} = time of the event from parent to child,
\item \code{node_area} = the range area in the event,
\item \code{event_type} = the type of event,
\item \code{source} = can be 'cladogenesis' or 'anagenesis'
}
}
\description{
Get data frame of transition states from Biogeographical Stochastic Mapping
}
