% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_getNodeRange_biogeobears.R
\name{get_node_range_BioGeoBEARS}
\alias{get_node_range_BioGeoBEARS}
\title{Get node ranges from BioGeoBEARS biome reconstruction model}
\usage{
get_node_range_BioGeoBEARS(
  BioGeoBEARS.data,
  phyllip.file,
  tree,
  max.range.size
)
}
\arguments{
\item{BioGeoBEARS.data}{An object containing the result of BioGeoBEARS reconstruction}

\item{phyllip.file}{Phylip file, the same used in BioGeoBEARS package}

\item{tree}{A phylogenetic tree}

\item{max.range.size}{Scalar indicating the maximum number of biomes in which each node can occupy}
}
\value{
Data frame with two columns, one indicating the node and other the ancestral biome range
}
\description{
This is an auxiliary function that can be with a ancestral area reconstruction
from BioGeoBEARS to obtain a data frame containing the ancestral area of occurrence (biome range)
for all ancestors (nodes)
}
\examples{
data(resDEC) # ancestral area reconstruction from biogeobears
data(akodon_newick)
extdata_dir = system.file("extdata", package="Herodotools")
fn = paste(extdata_dir, "/geo_area_akodon.data", sep="")
node_area <- get_node_range_BioGeoBEARS(resDEC,phyllip.file = fn, akodon_newick,  max.range.size = 4) # obtaining area for each node

}
