% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_gridfilter.R
\name{GridFilter}
\alias{GridFilter}
\title{Spatial Grid generator}
\usage{
GridFilter(shape, resol = 1, prop = 0)
}
\arguments{
\item{shape}{A raster object}

\item{resol}{Numeric value indicating the resolution of grid cells}

\item{prop}{Numeric value indicating the minimum proportion to be contained in a cell to build the grids}
}
\value{
A raster object containing the grids with specified values
}
\description{
Spatial Grid generator
}
