% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/insert_nodes.R
\name{insert_nodes}
\alias{insert_nodes}
\title{Insert internal nodes in the phylogenetic tree}
\usage{
insert_nodes(tree, inserts, node_area = NULL)
}
\arguments{
\item{tree}{An object of class 'phylo'.}

\item{inserts}{a data frame create with the function \code{\link[=get_insert_df]{get_insert_df()}}}

\item{node_area}{default NULL. One column data frame indicating the range area of each node (rows)}
}
\value{
a list with 2 elements:
\itemize{
\item \verb{$phylo}: a phylogenetic tree with added non-bifurcating nodes
\item \verb{$node_area}: a data frame with the node area for each added node. If the
argument \code{node_area} is not NULL, all the node area are included
in the data frame.
}
}
\description{
It adds non-bifurcating nodes in the branches of the phylogenetic
tree to represent events of ancestral character changes, specifically change
in range states from biogeographic ancestral reconstruction.
}
