% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_nodes_info_core.R
\name{nodes_info_core}
\alias{nodes_info_core}
\title{Auxiliary function to compute information of node path and dispersal path for each species}
\usage{
nodes_info_core(W, tree, ancestral.area, biogeo)
}
\arguments{
\item{W}{Numerical matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic tree in newick format}

\item{ancestral.area}{Single column data frame with nodes in rows and ancestral area/Ecoregion of occurrence in columns. If the reconstruction
comes from BioGeoBears this data frame can be obtained with (\code{\link{get_node_range_BioGeoBEARS}})}

\item{biogeo}{Data frame containing the information of the biome/area/ecoregion of each assemblage}
}
\value{

}
\description{
Auxiliary function to compute information of node path and dispersal path for each species
}
\examples{

}
