% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_organize_pd_decomp.R
\name{organize_pd_decomp}
\alias{organize_pd_decomp}
\title{Auxiliar to organize output from PD decomposition function}
\usage{
organize_pd_decomp(list1.decomp, list2.treetable)
}
\arguments{
\item{list1.decomp}{data frame with PD components}

\item{list2.treetable}{tibble with edge partition for each community}
}
\value{
A list object with decomposition of PD
}
\description{
Auxiliar to organize output from PD decomposition function
}
