% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PD_ada_multiple.R
\name{PD_ada}
\alias{PD_ada}
\title{Calculation of PD using ada reconstruction}
\usage{
PD_ada(ada.obj, thresold, comm)
}
\arguments{
\item{ada.obj}{}

\item{thresold}{}

\item{comm}{}
}
\description{
Calculation of PD using ada reconstruction
}
