% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_pd_decomposition_sbears.R
\name{PD_decomposition}
\alias{PD_decomposition}
\title{Phylogenetic diversity decomposition}
\usage{
PD_decomposition(comm, sbears.obj, phy, threshold, make.node.label = TRUE)
}
\arguments{
\item{comm}{a community matrix with sites in rows and species in columns}

\item{sbears.obj}{an object calculated from \code{\link{calc_sbears}}
containing ancestral community reconstruction}

\item{phy}{A phylo object containing phylogenetic relationships among species}

\item{threshold}{Scalar indicating the threshold used to consider the
presence of a species in a community}

\item{make.node.label}{Logical. If TRUE (the default) the nodes of the phylogeny
will be named with the letter N preceding the number of the node}
}
\value{
a list of length two.
\itemize{
\item \code{decomp_potential}: is a list with results of PD decomposition
considering the range area model reconstruction. This list contains:
\item \code{PD_decomposition}: A data frame with three colums.
\code{partition} is a character indicating the component of PD,
\code{value} is a numeric columns with respective PD compoent and
\code{community} is a character column with the name of the community

\if{html}{\out{<div class="sourceCode">}}\preformatted{ \\item \code{decomp_faith}: is a list with the same structure of  
     \code{decomp_potential}, but contains the results of PD decomposition
     considering only the current occurrence of lineages in communities
}\if{html}{\out{</div>}}

}
}
\description{
Phylogenetic diversity decomposition
}
\examples{
# phylogenetic tree
set.seed(42)
phylo <- ape::rcoal(n = 10)
phylo <- ape::makeNodeLabel(phy = phylo)

# community composition matrix
comm <- 
  matrix(sample(c(0, 1), size = 10*20, replace = TRUE),
         nrow = 20, 
         ncol = 10, 
         dimnames = list(paste("comm", 1:20), phylo$tip.label)
  )

# coordinates - this is necessary for "disperal_assembly" algorithm
xy_coords <- 
  matrix(runif(1:10), 
         nrow = nrow(comm),
         ncol = 2, 
         dimnames = list(rownames(comm), c("lng", "lat")
         )
  )

# running sbears with "single_site" algorithm
out_sbears <- 
  calc_sbears(x = comm,
              phy = phylo, 
              coords = xy_coords, 
              method = "single_site")

# sbears object can be used directly in PD_decomposition function

out_decomp <- 
  PD_decomposition(comm = comm,
                   sbears.obj = out_sbears, 
                   phy = phylo, 
                   threshold = 0.5)

# Data frame with results from PD decomposition
out_pd_decomposition <- out_decomp$decomp_potential$PD_decomposition
    
    
}
