% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{phy_sbears}
\alias{phy_sbears}
\title{Phylogeny of 260 Sigmodontinae species}
\format{
A phylo object containing 260 tips and 259 internal nodes
describing the phylogenetic relationship among 260 sigmodontine
species. The tree is rooted and include branch lengths
}
\source{
\url{https://zenodo.org/records/16933539}
}
\usage{
phy_sbears
}
\description{
Phylogeny of Sigmodontinae species
}
\keyword{datasets}
