% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_ada.R
\name{plot_ada}
\alias{plot_ada}
\title{Plotting ancestral area in space}
\usage{
plot_ada(
  ada.res,
  grid,
  coords,
  resolution,
  patterns = "all",
  color_palette = "SunsetDark",
  projection = "+proj=robin"
)
}
\arguments{
\item{ada.res}{An object from \code{\link{ada}} function}

\item{grid}{A spatial object containing the cells used to calculate ancestral
diversity distribution with ada}

\item{coords}{A two column matrix containing the values of
Longitude (first column) and Latitude (second column)}

\item{resolution}{Scalar informing the resolution of grids in grid object}

\item{patterns}{Character, a vector containing the names of the metrics to be
spatialized or "all" to create a map for all metrics. The  characters
allowed to be passed are "rich", "Nnodes", "PeakDiv",
"Skewness", "LowDistPeak", "HighDistPeak", "PeakRange". Default is "all"}

\item{color_palette}{Character indicating the palette to be used to plot spatial patterns. it can be one
of the option present in \link[rcartocolor]{scale_fill_carto_c} function. Default is "SunsetDark"}

\item{projection}{Character indicating the projection to be used in spatial plots.
Default is "+proj=robin"}
}
\value{
A list containing spatial plots for the metrics calculated in ada.res
}
\description{
Plotting ancestral area in space
}
\keyword{internal}
