% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_age_comm.R
\name{plot_age_arrival}
\alias{plot_age_arrival}
\title{Plot the mean arrival age of species in an assemblage based in an ancestral area reconstruction across geographic space}
\usage{
plot_age_arrival(
  age_arrival_comm,
  coords,
  shapefile,
  color_palette = "SunsetDark"
)
}
\arguments{
\item{age_arrival_comm}{A list of length two generated by
\code{\link{Herodotools::calc_age_arrival}} containing community-level
the mean arrival age of species in an assemblage based in an ancestral area reconstruction.}

\item{coords}{A rectangular object (it can be a matrix, data.frame or tibble)
containing two columns, one with longitude and another with latitude
coordinates in this order.}

\item{shapefile}{An \code{sf} object containing the geographic boundaries to
be overlaid on the raster map (e.g., continents, biogeographic regions, or
study area limits).}

\item{color_palette}{character string specifying the color palette to be used for
quantitative variables in \code{rcartocolor::scale_fill_carto_c}.}
}
\value{
A \code{ggplot2} object is returned.
}
\description{
Plot the mean arrival age of species in an assemblage based in an ancestral area reconstruction across geographic space
}
\examples{
\dontrun{

# Compute diversification metrics

akodon_age_comm <- Herodotools::calc_age_arrival(W = akodon_pa_tree,
                                         tree = akodon_newick,
                                         ancestral.area = node_area,
                                         biogeo = biogeo_area)
#Load data

site_xy <- akodon_sites |>
    dplyr::select(LONG, LAT)

coastline <- rnaturalearth::ne_coastline(returnclass = "sf")

coastline_crop <- sf::st_crop(coastline,
                          xmin = map_limits$x[1],
                          xmax = map_limits$x[2],
                          ymin = map_limits$y[1],
                          ymax = map_limits$y[2])


plot_age_arrival(age_arrival_comm = akodon_age_comm,
                 coords = site_xy,
                 shapefile = coastline_crop)

}

}
\author{
Maria Gabriela Junqueira and Gabriel Nakamura
}
