% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_evoregions.R
\name{plot_evoregions}
\alias{plot_evoregions}
\title{Plot Evoregions Classification}
\usage{
plot_evoregions(evoregions, coords, shapefile)
}
\arguments{
\item{evoregions}{A list generated by
\code{\link{calc_evoregions}}}

\item{coords}{A rectangular object (it can be a matrix, data.frame or tibble)
containing two columns, one with longitude and another with latitude
coordinates in this order.}

\item{shapefile}{An \code{sf} object containing the geographic boundaries to
be overlaid on the raster map (e.g., continents, biogeographic regions, or
study area limits).}
}
\value{
A \code{ggplot2} object is returned.
}
\description{
Plot Evoregions Classification
}
\examples{
\dontrun{

# Compute evoregions

akodon_regions <-  Herodotools::calc_evoregions(comm = akodon_pa_tree,
                                         phy = akodon_newick)

#' #Load data

site_xy <- akodon_sites |>
    dplyr::select(LONG, LAT)

coastline <- rnaturalearth::ne_coastline(returnclass = "sf")

coastline_crop <- sf::st_crop(coastline,
                          xmin = map_limits$x[1],
                          xmax = map_limits$x[2],
                          ymin = map_limits$y[1],
                          ymax = map_limits$y[2])

plot_evoregions(evoregions = akodon_regions,
                 coords = site_xy,
                 shapefile = coastline_crop)

}

}
\author{
Maria Gabriela Junqueira and Gabriel Nakamura
}
