% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_region_occ.R
\name{region_occ}
\alias{region_occ}
\title{Species occurrence in each evoregion}
\usage{
region_occ(comm, site.region)
}
\arguments{
\item{comm}{Occurrence data frame with species in colums and rows corresponding to assemblages}

\item{site.region}{a character vector indicating the membership of each assemblage to regions}
}
\value{
An occurrence data frame with species as rownames and regions as columns
}
\description{
Auxiliary function to produce an occurrence data frame of species in each region. This object can be
with \code{\link{tipranges_to_BioGeoBEARS}} function to produce a phyllip file needed in BioGeoBEARS
ancestral area reconstruction
}
\examples{
data(akodon.pa.tree)
data(regions)
site_evoregion <- regions$Cluster_Evoregions
region_occ(comm=akodon.pa.tree,site.region=site_evoregion)
}
\seealso{
\code{\link{tipranges_to_BioGeoBEARS}}
}
\author{
Gabriel Nakamura and Arthur Rodrigues
}
