% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{regions}
\alias{regions}
\title{Output from \code{evoregion} function with classification of 732 assemblages}
\format{
An object of class \code{evoregion} of length 2.
}
\usage{
regions
}
\description{
Object containing a list with two objects:
\describe{
\item{$PCPS}{A list containing eigenvectors from PCPS analysis}
\item{$Cluster_Evoregions}{A vector containing the groups of each assemblage}
}
}
\keyword{datasets}
