% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{resDEC}
\alias{resDEC}
\title{Output from BioGeoBEARS}
\format{
A list obtained from BioGeoBEARS with Akodon species
}
\usage{
resDEC
}
\description{
Object from a ancestral reconstruction model with BioGeoBEARS
}
\keyword{datasets}
