% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{rodent_phylo}
\alias{rodent_phylo}
\title{Phylogenetic tree with 285 species}
\format{
Newick phylogenetic tree
}
\usage{
rodent_phylo
}
\description{
Dated (my) phylogenetic tree in newick format containing 285 of sigmodontine species
}
\keyword{datasets}
