% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{site_region}
\alias{site_region}
\title{Classification of communities in evoregions}
\format{
A vector of length 732 and five levels
}
\source{
\url{https://zenodo.org/records/16933539}
}
\usage{
site_region
}
\description{
Classification of communities in evoregions for akodon genus in
the neotropics with factors indicating the number of evoregions
}
\keyword{datasets}
