% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_association.R
\name{spp_association_evoreg}
\alias{spp_association_evoreg}
\title{Quantifying species association with evoregions}
\usage{
spp_association_evoreg(comm, evo.vectors, spp.association = 0.7)
}
\arguments{
\item{comm}{A matrix with assemblages in rows and species in columns}

\item{evo.vectors}{A list returned from evoregions function}

\item{spp.association}{A scalar indicating the level of association that will be evaluated.
Default is 0.7}
}
\value{
A matrix with species association value for each assemblage
}
\description{
Quantifying species association with evoregions
}
\examples{
\dontrun{
data(akodon.pa.tree) # occurrence data 
data(akodon.newick) # phylogenetic tree
regions <- evoregions(comm = akodon.pa.tree, phy = akodon.newick) # compute evoregions
spp_association_evoreg(comm = akodon.pa.tree, evo.vectors = regions, spp.association = 0.7) # calculate species association
}

}
