% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/std_name_evoregion.R
\name{std_name_evoregion}
\alias{std_name_evoregion}
\title{Standardize the evoregion names after 'adegenet::find.clusters'}
\usage{
std_name_evoregion(clust_evoreg)
}
\arguments{
\item{clust_evoreg}{group vector from the 'adegenet::find.clusters'}
}
\value{
factor vector with the groups
}
\description{
Although the results of 'find.cluster' are consistent when function is runned multiple times, the order of the names of the groups are random.
Here, the function reorder the name of the groups following the sequence the appear in the vector.
}
\details{
The function was designed to be used within the function \code{`calc_evoregions`}
}
