% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/transition_rates_function.R
\name{tip.based.trait.evo}
\alias{tip.based.trait.evo}
\title{Tip-based metrics of trait evolution}
\usage{
tip.based.trait.evo(
  tree,
  trait,
  nsim = 1,
  method = c("transition_rates", "last_transition_time", "stasis_time")
)
}
\arguments{
\item{tree}{A phylogenetic tree as an object of class "phylo"}

\item{trait}{A named vector containing the tip states for a discretely valued character. The names must match the tip labels of the tree.}

\item{nsim}{Number of simulations to stochastic character mapping.}

\item{method}{Tip-based metric, partial match to "transition_rates", "last_transition_time" and "stasis_time".}
}
\value{
A list (length equal to nsim) with tip-based metrics estimated per species.
}
\description{
Tip-based metrics of trait evolution
}
\author{
André Luza and Vanderlei Debastiani
}
