% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tipranges_to_BioGeoBEARS.R
\name{tipranges_to_BioGeoBEARS}
\alias{tipranges_to_BioGeoBEARS}
\title{Generates lagrangePHYLIP file for using in BioGeoBEARS}
\usage{
tipranges_to_BioGeoBEARS(
  tip_range,
  filename = "lagrange_area_data_file.data",
  areanames = NULL
)
}
\arguments{
\item{tip_range}{data frame. Tip names in rows, geographical areas in columns.
Please provide the tip names as \code{row.names}. See example.}

\item{filename}{filename to store the data. By default it is
\code{"lagrange_area_data_file.data"}.}

\item{areanames}{the names of the areas. By defaut the areanames are generated
as sequencial upper case letters.}
}
\description{
It prepares and save the file to be used as geographical areas in \code{BioGeoBEARS}
scripts. It was based on \code{save_tipranges_to_LagrangePHYLIP} function of
\code{BioGEOBEARS} package.
}
\examples{

\dontrun{
set.seed(082622)
geo.areas <- data.frame(
  a = rbinom(5, 1, 0.5),
  b = rbinom(5, 1, 0.5),
  c = rbinom(5, 1, 0.5)
)

# set the row names
row.names(geo.areas) <- paste0("tip_", 1:nrow(geo.areas))

# it saves the data at 'filename' path.
tipranges_to_BioGeoBEARS(
  geo.areas, 
  filename = "lagrange_area_data_file.data", 
  areanames = NULL)
)

# file content must be like this:
# 5	3	(A B C)
# tip_1	111
# tip_2	001
# tip_3	010
# tip_4	101
# tip_5	010
 }
 
}
