% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{toy_treeEx}
\alias{toy_treeEx}
\title{Hypothetical phylogenetic tree containing five species}
\format{
A newick phylogenetic tree with 5 species
}
\usage{
toy_treeEx
}
\description{
This is a made up data that simulates a phylogenetic tree in newick format. This tree contains
five species and the annotations in each node corresponds to the occurrence area of ancestors
}
\keyword{datasets}
