% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{trait}
\alias{trait}
\title{Trait on foraging strata for small rodent}
\format{
character vector with three categories:
\describe{
\item{Ar}{Arboreal}
\item{G}{Ground level}
\item{S}{Scansorial}
}
}
\usage{
trait
}
\description{
Character vector indicating the foraging strata for 214 species of small rodents
}
\keyword{datasets}
