% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_ada.R
\name{ada}
\alias{ada}
\title{Ancestral Diversity Analysis}
\usage{
ada(
  x,
  phy,
  sp.bin = "Sturges",
  marginal = FALSE,
  lik.threshold = FALSE,
  threshold = 0.7,
  compute.fields = F
)
}
\arguments{
\item{x}{Community occurrence matrix. Rows are sites and columns are species}

\item{phy}{Phylogenetic tree}

\item{sp.bin}{Character indicating the methods to be used to compute the number of time slices in which metrics will be computed. Default is "Sturges"}

\item{marginal}{Logical indicating}

\item{lik.threshold}{Logical indicating if some threshold value will be used to select the likelihood values of ancestor species in a site. Default is TRUE}

\item{threshold}{Scalar indicating the threshold value used to select the likelihood of a given species be present at a site}

\item{compute.fields}{Logical, indicates if phylogenetic fields will be computed. Default is FALSE}
}
\value{
A list of size eight containing the following objects
\item{Phylogeny}{Phylogenetic tree} \item{Root.Age}{Numeric containing root age of the tree used}
\item{Per.node.ancestral.area}{Matrix containing the ancestors (nodes) for each species }
\item{Diversity.Through.Time}{Age of each node in the phylogeny}
\item{Cell.Metric}{Matrix containing the assemblage level metrics for richness and ancestral diversity}
}
\description{
This function computes different assemblage level
metrics using an ancestral community reconstruction
}
\details{
ada function calculates an ancestral assemblage reconstruction for each assemblage
in a matrix using for this reconstruction \link[ape]{ace} function from ape package
}
\examples{

}
