% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliation_evoreg.R
\name{afilliation.evoreg}
\alias{afilliation.evoreg}
\title{Affiliation values for assemblages according to phylogenetic turnover}
\usage{
afilliation.evoreg(evo.vectors, n.groups = 4, method = "euclidean")
}
\arguments{
\item{evo.vectors}{An object returned from evoregions function}

\item{n.groups}{Scalar indicating the number of groups that will be evaluated for affiliation}

\item{method}{Character indicating the method used to compute the distance in fuzzy matrix}
}
\value{
A list with two elements containing affiliation values and species coordinates
}
\description{
Affiliation values for assemblages according to phylogenetic turnover
}
