% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/affiliation_evoreg.R
\name{afilliation.evoreg}
\alias{afilliation.evoreg}
\title{Affiliation values for assemblages according to phylogenetic turnover}
\usage{
afilliation.evoreg(evo.vectors, method = "euclidean")
}
\arguments{
\item{evo.vectors}{An object returned from \code{\link{evoregions}} function}

\item{method}{Character indicating the method used to compute the distance in fuzzy matrix}

\item{n.groups}{Scalar indicating the number of groups that will be evaluated for affiliation}
}
\value{
A list with two matrix, one containing affiliation values and the group in which each cell
is classified and the other containing cell coordinates
}
\description{
Affiliation values for assemblages according to phylogenetic turnover
}
