% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calc_ed.R
\name{calc_ed}
\alias{calc_ed}
\title{Calculate a evolutionary distinctiveness (ED) based on biogeographical model}
\usage{
calc_ed(
  tree,
  ancestral.area,
  current.area = NULL,
  type = c("equal.splits", "fair.proportion")
)
}
\arguments{
\item{tree}{Phylogenetic tree of class 'phylo'.}

\item{ancestral.area}{One column data frame indicating the Ecoregion of occurrence of each node (rows)}

\item{current.area}{character. A single area that will represent the current area of all tips in the phylogeny.}

\item{type}{Character indicating the type of evolutionary distinctiveness metric to be used.
It can be "equal.splits" or "fair.proportion", "equal-splits" is the default.}
}
\value{
a named numeric vector with values for all tips in the phylogeny
}
\description{
For each tip, the function calculates the ED based on the current area and the
ancestral area based on a biogeographical model. ED is calculated for the connecting
branched from tip to internal node that are in the same area as the current area.
This means that the evolutionary distinctiveness occurred in situ (same biogeographical region).
If current area is not provided, the function gives same result as 'picante::evol.distinct()'
}
\examples{

# example for calc_ed
# generate simlutated data ----
set.seed(4523)
tree_sim <- ape::rcoal(5)

# Create node area table (only bifurcating nodes)
node_area <- data.frame(
    area = c("A", "A", "BC", "AD"),
    row.names = paste0("N", 6:9)
    )
    
# create insertion of nodes to simulate for change in area in the branch 
# Use ggtree to identify insertion points 

gdata <- ggtree::ggtree(tree_sim)$data
ins_data <- gdata \%>\% dplyr::filter(node \%in\% c(3, 9, 8))
ins_data <- tibble::add_row(ins_data, ins_data[2, ]) \%>\% arrange(node)

# Define inserted nodes with areas
inserts <- tibble::tibble(
  parent = ins_data$parent,
  child = ins_data$node,
  event_time = c(0.2, 0.5, 0.9, .3),
  node_area = c("AC", "AB", "BC", "D")
  )
  
# Insert nodes 
result <- insert_nodes(tree_sim, inserts, node_area = node_area)
tree_out <- result$phylo
node_area_out <- result$node_area


# Compute Evolutionary Distinctiveness --------

# ED total (same results as picante::evol.distinct())
ed_total <- calc_ed(  
  tree_out, 
  node_area = node_area_out, 
  type = "equal.splits")
  
# ED partial. Account for in situ distinctiveness only

ed_partial_A <- calc_ed(
  tree_out,
  node_area = node_area_out, 
  current_area = "A", 
  type = "equal.splits")
  
ed_partial_D <- calc_ed(
  tree_out,
  node_area = node_area_out, 
  current_area = "D", 
  type = "equal.splits")
  
data.frame(
  species = names(ed_total), 
  ed_total,
  ed_partial_A,
  ed_partial_D
  )
  
# tree visualization
tree_viz <- tree_out
tree_viz$node.label <- node_area_out$area

plot(tree_viz, show.node.label = T)
nodelabels(adj = 1.2)



}
