% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_model_based_diversificationJetz.R
\name{calc_insitu_diversification}
\alias{calc_insitu_diversification}
\title{Tip-based and model-based diversification metrics}
\usage{
calc_insitu_diversification(
  W,
  tree,
  ancestral.area,
  biogeo,
  diversification = "jetz",
  type = "equal.splits"
)
}
\arguments{
\item{W}{Assemblage occurrence matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic tree in newick format}

\item{ancestral.area}{One column data frame indicating the Ecoregion of occurrence of each node (rows)}

\item{biogeo}{Data frame with one column indicating the Ecoregion of each assemblage}

\item{diversification}{Character indicating the tip-based diversification metric to be calculated. Default is "jetz"}

\item{type}{Character indicating the type of calculation to be used in the calculation of tip-based metric, "equal-splits" is the default}
}
\value{
A list with the following objects:
\itemize{
\item{Jetz_per_spp}{Jetz tip-based metric of diversification for each species}
\item{Jetz_species_site}{Matrix containing
the values of Jetz tip-based diversification metric for each assemblage}
\item{Jetz_harmonic_mean_site}{Harmonic mean of Jetz tip-based
diversification metric}
\item{model_based_Jetz_species_sites}{Matrix with model-based Jetz metric for each species
in each assemblage}
\item{model_based_Jetz_harmonic_mean_site}{Harmonic mean of model-based Jetz diversification}
}
}
\description{
Tip-based and model-based diversification metrics
}
\details{
model-based measure of diversification is calculated as the proportion of total tip-based diversification
metric corresponded only to the branches of a given species that diversified in the current area of occurrence of that species
\deqn{Jetz_total \times (ED_div/ED_total)}
}
\examples{
\dontrun{
W_toy <- matrix(c(0, 1, 1, 0, 1, 
    1, 0, 1, 0, 1, 
    0, 0, 1, 0, 0),nrow= 3,
    ncol= 5,dimnames=list(c("Comm 1", "Comm 2", "Comm 3"),
                          c(paste("s", 1:5, sep=""))))
data("toy_treeEx")
toy_treeEx                   
biogeo_toy <- data.frame(Ecoregion= c("A", "B", "C"))
ancestral_area_toy <- data.frame(state= c("ABC", "B", "C", "ABC"))
plot(toy_treeEx, show.node.label = TRUE)
calc_insitu_diversification(W = W_toy,
                            tree = toy_treeEx, 
                            ancestral.area = ancestral_area_toy,
                             biogeo = biogeo_toy)
} 

}
\author{
Gabriel Nakamura \href{mailto:gabriel.nakamura.souza@gmail.com}{gabriel.nakamura.souza@gmail.com} and Arthur V Rodrigues
}
