% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_sbears.R
\name{calc_sbears}
\alias{calc_sbears}
\title{Site-Based Estimation of Ancestral Range of Species}
\usage{
calc_sbears(
  x,
  phy,
  coords,
  method = c("single_site", "disp_assembly"),
  w_slope = 5,
  min_disp_prob = 0.8,
  compute.node.by.sites = TRUE,
  make.node.label = TRUE
)
}
\arguments{
\item{x}{Matrix object with community/assemblage composition.
Rows represent sites and columns represents species.}

\item{phy}{A phylogenetic tree of \code{phylo} class object.}

\item{coords}{A rectangular object (it can be a matrix, data.frame or tibble)
containing two columns, one with longitude and another with latitude
coordinates in this order.}

\item{method}{character indicating how ancestral range probabilities are computed.
The options are "single_site" or "disp_assembly"}

\item{w_slope}{A scalar representing the slope of the dispersal kernel function.}

\item{compute.node.by.sites}{Logical, TRUE (default) computes a matrix of node occurrence by site.}

\item{make.node.label}{Logical, if TRUE (default) the nodes of the phylogeny will be named as the letter "N" preceding node number}
}
\value{
a list with three elements. reconstruction is the result of ancestral
.   area reconstruction; phylogeny is the matrix containing
the occurrence of nodes in sites and joint.phylo.obs is the joint occurrence
of nodes and species in phylogenetic tree.
}
\description{
Site-Based Estimation of Ancestral Range of Species
}
\details{
SBEARS (Site-Based Estimation of Ancestral Range of Species) is a method for
ancestral state reconstruction of species geographic ranges. It operates at
fine spatial resolution and does not require predefined discrete biogeographic
areas. The method estimates ancestral ranges under a maximum likelihood
framework, using site-level occurrence information.
\itemize{
\item the method used to compute ancestral range probabilities can be
\code{single_site} or \code{disp_assembly}. For both methods SBEARS use the function
\code{\link[phytools]{fastAnc}} to reconstruct the the ancestral ranges.
In \code{single_site} each site is taken
as a binary categorical trait representing the presence (state 1) or
the absence (state 0) of every species in the site, which is reconstructed
along the nodes of the phylogeny.
In \code{disp_assembly} option the single site reconstruction is
carried out, producing a matrix describing standardized probabilities
of occurrence of each node at each site. The difference here is that
an additional step is performed to compute the probability of a cell
being colonized by an ancestral species originating from a focal cell.
The probability of colonization decreases increasing the distance from
the focal cell. This probability of an ancestral species to disperse
from a cell to any other can be defined using a dispersal kernel function
proposed by Gravel et al (2006).
}
}
\examples{

data("akodon_newick")
data("akodon_sites")


site_xy <- akodon_sites \%>\% 
  dplyr::select(LONG, LAT) 

akodon_pa <- akodon_sites \%>\% 
  dplyr::select(-LONG, -LAT)
 
akd <- picante::match.phylo.comm(akodon_newick, akodon_pa)

akodon_sbears <- calc_sbears(x = ak$comm, phy = ak$phy, coords = site_xy)

# Visualize root node area

sbears_df <- cbind(site_xy, akodon_sbears$PD_nodes_by_sites)

ggplot(sbears_df) +
  geom_tile(aes(x = LONG, y = LAT, fill = Node1))


}
\references{
Gravel D., Canhan C.D., Beaudet M. and Messier C. Reconciling
niche and neutrality: the continuum hypothesis. 2006. Ecology Letters
\doi{10.1111/j.1461-0248.2006.00884.x}
}
