% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spp_association.R
\name{calc_spp_association_evoreg}
\alias{calc_spp_association_evoreg}
\title{Quantifying species association with evoregions}
\usage{
calc_spp_association_evoreg(comm, evo.vectors, spp.association = 0.7)
}
\arguments{
\item{comm}{A matrix with assemblages in rows and species in columns}

\item{evo.vectors}{A list returned from \code{calc_evoregions} function}

\item{spp.association}{A scalar indicating the level of association that will be evaluated.
Default is 0.7}
}
\value{
A matrix with species association value for each assemblage
}
\description{
Quantifying species association with evoregions
}
\examples{
\dontrun{
data(akodon_sites) # occurrence data 
akodon_pa <- akodon_sites \%>\% dplyr::select(-LONG, -LAT)
data(akodon_newick) # phylogenetic tree
spp_in_tree <- names(akodon_pa) \%in\% akodon_newick$tip.label
akodon_pa_tree <- akodon_pa[, spp_in_tree]
regions <- calc_evoregions(comm = akodon_pa, phy = akodon_newick) # compute evoregions
calc_spp_association_evoreg(comm = akodon_pa_tree, 
                       evo.vectors = regions, 
                       spp.association = 0.7) # calculate species association
}

}
