% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_comp_ada_nodes_sites.R
\name{comp_ada_nodes_sites}
\alias{comp_ada_nodes_sites}
\title{Compute a matrix of node occurrences in each site}
\usage{
comp_ada_nodes_sites(phy, comm, long = FALSE)
}
\arguments{
\item{phy}{a phylogenetic tree}

\item{comm}{a community matrix}
}
\value{
a data frame or matrix in dense or long format. Default is dense format (sites x nodes)
}
\description{
Compute a matrix of node occurrences in each site
}
