% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aux_kernel_gravel_density_dist_decay.R
\name{comp_kernel_gravel}
\alias{comp_kernel_gravel}
\title{Occupancy probability of each node in each site considering a kernel density funcion}
\usage{
comp_kernel_gravel(x, phy, coords, w_slope, min_disp_prob)
}
\arguments{
\item{x}{Node by communities matrix}

\item{phy}{phylo object}

\item{coords}{Data frame with two columns containing spatial coordinate of
communities}

\item{w_slope}{A scalar representing the slope of the dispersal kernel function.
This is the same as passed to \code{calc_sbears} function}

\item{min_disp_prob}{A scalar between 0 and 1 representing the minimum
probability of receiving an immigrant node from a focal cell. Same as
passed to \code{calc_sbears} function}
}
\value{
A list with length equal the number of sites. Each element of the list
contains a matrix with nodes in rows and sites in columns and the probability
of occurrence of each ancestral node in each site.
}
\description{
Occupancy probability of each node in each site considering a kernel density funcion
}
