% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_description.R
\docType{data}
\name{coords_sbears}
\alias{coords_sbears}
\title{Spatial coordinates of sites}
\format{
A data frame object with 46 rows and 2 columns:
\describe{
\item{longitude}{Numeric. Longitude of each site in decimal degrees}
\item{latitude}{Numeric. Latitude of each site in decimal degrees}
}
}
\source{
\url{https://zenodo.org/records/16933539}
}
\usage{
coords_sbears
}
\description{
Spatial coordinates containing longitute and latitude of 46
sites
}
\keyword{datasets}
