% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_dispersal_from.R
\name{dispersal_from}
\alias{dispersal_from}
\title{Proportional contribution of each region to the species of another region}
\usage{
dispersal_from(W, tree, ancestral.area, biogeo)
}
\arguments{
\item{W}{Occurrence matrix, rows are assemblages and columns are species}

\item{tree}{Phylogenetic tree in newick format}

\item{ancestral.area}{One column data frame with nodes in rows and one column indicating the occurrence (biome/ecoregion) area of nodes}

\item{biogeo}{One column data frame with assemblages in rows and their respective biome/ecoregion}
}
\value{

}
\description{
Proportional contribution of each region to the species of another region
}
\author{
Arthur V Rodrigues \href{mailto:rodrigues.arthur.v@gmail.com}{rodrigues.arthur.v@gmail.com}
}
