% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/DivB_metrics.R
\name{DivB_metrics}
\alias{DivB_metrics}
\title{Diversification Based (DivB) Metrics for Community Phylogenetics}
\usage{
DivB_metrics(
  W,
  tree,
  ancestral.area,
  biogeo,
  diversification = c("jetz", "freck"),
  PD = TRUE,
  PE = TRUE,
  age.arrival = TRUE,
  age.no.ancestor = NA,
  dispersal.from = TRUE,
  ED.type = "equal.splits"
)
}
\arguments{
\item{W}{community matrix (sites in the rows and species in the columns)}

\item{tree}{phylogenetic tree}

\item{ancestral.area}{biogeographic area for ancestral. Must be a single
column data.frame in which rows are nodes in the tree. Each area must be a
single letter that could be combined or not (eg. "A", "AB", "B" ).}

\item{biogeo}{single column data frame with biogeographic area for each site. The areas must match with the same code provided in
ancestral.area argument}

\item{diversification}{A character indicating the Diversification index that will be computed.
Defaut = c("jetz", "freck").}

\item{PD}{Logical. calculate phylogenetic diversity and its diversification based version (PDlocal). Defaut = TRUE.}

\item{PE}{Logical. calculate phylogenetic endemism and its diversification based version (Db-PE). Defaut = TRUE. See details for more
information}

\item{age.arrival}{logical. Calculate mean age of arrival of species in a given site. Default = TRUE.}

\item{age.no.ancestor}{How to deal with age arrival for species in which the
late ancestor wasn't in the site's biogeographic area. Options: NA,
numeric(), or 'half.edge'. Default = NA.  See Details. (not implemented yet)}

\item{dispersal.from}{logical. Calculate frequency of dispersal from each
biogeographic area of ancestors for each site}
}
\description{
Compute model-based diversification metrics (Jetz et al 2008 and Freckleton et al 2008),
diversity metrics (Phylogenenic diversity and phylogenetic endemism),
age of ancestor arrival in assemblages, and frequency of dispersal
from one biogeographic area to another
}
