% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoregions.R
\name{evoregions}
\alias{evoregions}
\title{Defining groups based on phylogenetic turnover}
\usage{
evoregions(
  comm,
  coords,
  phy,
  method.dist = "bray",
  tresh.dist = 0.05,
  method.clust = "kmeans",
  stat.clust = "BIC",
  n.iter.clust = 1e+07,
  criterion.clust = "diffNgroup",
  max.n.clust = 10
)
}
\arguments{
\item{comm}{Community occurrence matrix. Assemblages are rows and species are columns}

\item{coords}{Numerical matrix with coordinates for assemblages in comm}

\item{phy}{Newick object with phylogenetic tree containing the species in comm}

\item{method.dist}{Character with the method to be used to compute phyogenetic distances.
Default is "bray"}

\item{tresh.dist}{A scalar informing the distance treshold value to select eigenvectors.
Defalt is 0.05 (eigenvector with at least 5\% of variation)}

\item{method.clust}{Character indicating the grouping algorithm to be used in cluster analysis.
Defalt is "kmeans"}

\item{stat.clust}{Character}

\item{n.iter.clust}{Scalar indicating the number of}

\item{criterion.clust}{Character}

\item{max.n.clust}{Scalar indicating the maximum number of clusters to be returned}
}
\value{
A adgenet object containing the clusters
}
\description{
Defining groups based on phylogenetic turnover
}
