% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/evoregions.R
\name{evoregions}
\alias{evoregions}
\title{Defining groups based on phylogenetic turnover}
\usage{
evoregions(
  comm,
  phy,
  max.n.clust = NULL,
  max.n.clust.method = "elbow",
  method.dist = "bray",
  tresh.dist = 0.05,
  method.clust = "kmeans",
  stat.clust = "BIC",
  n.iter.clust = 1e+07,
  criterion.clust = "diffNgroup"
)
}
\arguments{
\item{comm}{Species occurrence matrix. Assemblages are rows and species are columns}

\item{phy}{Newick object with phylogenetic tree containing the species in comm}

\item{max.n.clust}{Integer value to be used in \code{\link{find.clusters}}.
Indicates the maximum number of clusters to be tried.}

\item{max.n.clust.method}{}

\item{method.dist}{Character. The method to be used to compute phyogenetic distances among assemblages.
Default is "bray"}

\item{tresh.dist}{A scalar informing the threshold value to select eigenvectors based on the amount of variation of each eigenvector.
Default is 0.05 (eigenvector with at least 5\% of variation)}

\item{method.clust}{Character indicating the grouping algorithm to be used in cluster analysis.
Defalt is "kmeans"}

\item{stat.clust}{Character to be used in \code{\link{find.clusters}} indicating the statistic to be computed for each model. Can be
"BIC" (default), "AIC" or "WSS".}

\item{n.iter.clust}{Integer to be used in \code{\link{find.clusters}}
function of adegenet package to indicate the number of iterations to be used in each run of K-means algorithm}

\item{criterion.clust}{Character indicating the type of method to be used to define the maximum number of clusters. Default option
is "elbow" method, implemented in \code{\link{find.clusters}} of phyloregion package.}
}
\value{
A list of length four containing
\item{PCPS_vectors}
A matrix with PCPS vectors
\item{Cluster_Evoregion}
A vector indicating the group in which each assemblage was classified
\item{Coordinates}
A matrix with spatial coordinates of assemblages
\item{Matrix_P}
Phylogenetic composition matrix obtained from \code{\link[=matrix.p]{matrix.p()}} function from SYNCSA package
\item{Total_vec_var}
A numeric vector with the amount of variation in the PCPS vectors selected accordingly with the threshold argument
}
\description{
Defining groups based on phylogenetic turnover
}
\examples{

}
\seealso{
\code{\link[=find.max.nclust]{find.max.nclust()}} to decide the maximum number of clusters to be used
}
