% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_find_maxn_clust.R
\name{find_max_nclust}
\alias{find_max_nclust}
\title{Estimate the maximum number of groups in DAPC analysis}
\usage{
find_max_nclust(
  x,
  threshold,
  max.nclust,
  nperm = 100,
  method = "kmeans",
  stat = "BIC",
  criterion = "diffNgroup",
  subset = 100,
  confidence.level = c(0.7, 0.8, 0.9, 0.95, 0.99)
)
}
\arguments{
\item{x}{A data.frame or matrix object containing eigenvectors by sites.}

\item{threshold}{Scalar. The number of eigenvectors used to perform classification.}

\item{max.nclust}{A vector containing values of the maximum number of groups to be evaluated.}

\item{nperm}{Scalar. Number of times classification will be performed.}

\item{method}{Character, one of c("kmeans","ward"). This will be used in \code{find.clusters} function.
See \code{\link{find.clusters}} of adegenet package. Default is "kmeans"}

\item{stat}{Character, one of c("BIC", "AIC", or "WSS"). This will be used in \code{find.clusters} function.
See \code{\link{find.clusters}} of adegenet package. Default is "BIC".}

\item{criterion}{Character one of c("diffNgroup", "min","goesup", "smoothNgoesup", or "goodfit").
This will be used in \code{find.clusters} function. Default is "diffNgroup".
See \code{\link{find.clusters}} of adegenet package.}

\item{subset}{Scalar. The number of cells used in the analysis.
It is particularly important whenever the total number of cells is large (> 1000).}

\item{confidence.level}{A vector containing values with threshold confidence
level used to estimate congruence in the classification pattern.}
}
\value{
Matrix containing congruence values ranging between 0-1 for each max.nclust
value (see Arguments) and confidence level.
}
\description{
Estimate the maximum number of groups in DAPC analysis
}
\details{
This function can be used to find the maximum number of clusters
that maximizes the congruence of the grouping procedure. Basically it
consists in repeat DAPC analysis multiple times using a set of candidate
maximum values (indicated in the argument \code{max.nclust})
}
\examples{
\dontrun{
data(regions)
evovectors <- regions$PCPS$vectors # eigenvectors by site
find_max_nclust(x = evovectors, threshold = 3, max.nclust = 10)
}
}
