% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_organize_pd_decomp.R
\name{organize_pd_decomp}
\alias{organize_pd_decomp}
\title{Organize output from PD decomposition}
\usage{
organize_pd_decomp(list1.decomp, list2.treetable)
}
\arguments{
\item{list1.decomp}{data frame with PD components}

\item{list2.treetable}{tibble with edge partition for each community}
}
\description{
Organize output from PD decomposition
}
