% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_partition_pd.R
\name{partition_pd}
\alias{partition_pd}
\title{Internal function used to calculate PD components}
\usage{
partition_pd(data, comm, phy)
}
\arguments{
\item{data}{a tibble containing the phylogenetic object}

\item{comm}{community matrix}

\item{phy}{newick phylogenetic object}
}
\description{
Internal function used to calculate PD components
}
