% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PD_decomposition.R
\name{PD_decomposition}
\alias{PD_decomposition}
\title{Phylogenetic diversity decomposition}
\usage{
PD_decomposition(comm, ada.obj, phy, threshold)
}
\arguments{
\item{comm}{a community matrix with sites in rows and species in columns}

\item{ada.obj}{an object calculated from \code{\link{ada_core}} containing community reconstruction}

\item{phy}{a newick object containing phylogenetic relationships among species}

\item{threshold}{a scalar indicating the threshold used to consider the presence of a species in a community}
}
\value{
a data frame with three columns. partition indicates the PD component, value is the values of partitions, community is the name of community
}
\description{
Phylogenetic diversity decomposition
}
