% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_PD_decomposition.R
\name{PD_decomposition}
\alias{PD_decomposition}
\title{Phylogenetic diversity decomposition}
\usage{
PD_decomposition(comm, ada.obj, phy, threshold)
}
\arguments{
\item{comm}{a community matrix with sites in rows and species in columns}

\item{ada.obj}{an object calculated from \code{\link{ada_core}} containing community reconstruction}

\item{phy}{a newick object containing phylogenetic relationships among species}

\item{threshold}{a scalar indicating the threshold used to consider the presence of a species in a community}
}
\value{
a list of length two. The element of the list named decomposition is a data frame with decomposition values of PD for all communities.
The element of the list named tree_table_potential is a tibble with node information for the potential tree for each community
}
\description{
Phylogenetic diversity decomposition
}
