% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_plot_dispersal.R
\name{plot_dispersal_from}
\alias{plot_dispersal_from}
\title{Plot historical dispersal contribution from source areas}
\usage{
plot_dispersal_from(
  dispersal_from,
  coords,
  shapefile,
  area_cols = NULL,
  area_col = NULL,
  color_palette = "SunsetDark",
  facet = FALSE
)
}
\arguments{
\item{dispersal_from}{A data.frame generated by \link[Herodotools]{calc_dispersal_from}  where each column represents
dispersal contribution from a source area, and each row corresponds
to a spatial cell}

\item{coords}{A rectangular object (it can be a matrix, data.frame or tibble)
containing two columns, one with longitude and another with latitude
coordinates in this order.}

\item{shapefile}{An \code{sf} object containing the geographic boundaries to
be overlaid on the raster map (e.g., continents, biogeographic regions, or
study area limits).}

\item{area_cols}{Integer vector indicating which columns of
\code{dispersal_from} should be used (plural).
If \code{NULL} (default), all columns are used.}

\item{area_col}{Integer indicating a single column of \code{dispersal_from}
to be plotted (singular). This argument is only used when
\code{facet = FALSE} and must be one of the values provided in
\code{area_cols}.}

\item{color_palette}{character string specifying the color palette to be used for
quantitative variables in \code{rcartocolor::scale_fill_carto_c}. Ignored
for \code{"DR_prop"}, which uses a fixed palette and scale. Default is "SunsetDark"}

\item{facet}{Logical. If \code{TRUE}, all areas defined in
\code{area_cols} are plotted using \code{facet_wrap()}.
If \code{FALSE}, a single area defined by \code{area_col} is plotted.}
}
\value{
A \code{ggplot} object.
}
\description{
This function plots spatial patterns of historical dispersal contribution from
one or multiple source areas. The user can visualize a single area
or multiple areas simultaneously using \code{facet_wrap()}.
}
\details{
When multiple areas are plotted, the order of the facets follows
the order of the columns provided in \code{area_cols}, not alphabetical order.

The function assumes that \code{dispersal_from} and \code{coords}
have the same number of rows and are ordered consistently.

All area selections are performed using numeric column indices,
ensuring an unambiguous mapping to the original columns of
\code{dispersal_from}.
}
\examples{
\dontrun{
# Compute diversification metrics

data("akodon_sites") # site in rows and species and coordinates in columns
data("akodon_newick") # phylogeny

# pre computed evoregion
load(file = system.file("extdata", "regions_results.RData", 
                        package = "Herodotools")) 
site_region <- regions$Cluster_Evoregions # evoregions

# subseting longitude and latitude in a separate dataframe
site_xy <- akodon_sites \%>\% 
  dplyr::select(LONG, LAT) 

# subsetting species presence in a separate dataframe
akodon_pa <- akodon_sites \%>\% 
  dplyr::select(-LONG, -LAT)

# joining classification with site coordinates
evoregion_df <- data.frame(
  site_xy, 
  site_region
)

# bioregion/ecoregion/evoregion of each site
biogeo_area <- data.frame(biogeo = chartr("12345", 
                                          "ABCDE", 
                                          evoregion_df$site_region)) 

# pre computed biogeobears model 
load(file = system.file("extdata", 
                        "resDEC_akodon.RData",
                        package = "Herodotools")) 

# subseting occurrence matrix keeping only species in the phylogenetic tree
spp_in_tree <- names(akodon_pa) \%in\% akodon_newick$tip.label
akodon_pa_tree <- akodon_pa[, spp_in_tree]


# getting the ancestral range area for each node 
node_area <- 
  Herodotools::get_node_range_BioGeoBEARS(
    resDEC,
    phyllip.file = here::here("inst", "extdata", "geo_area_akodon.data"),
    akodon_newick,
    max.range.size = 4 
  )

# Compute dispersal
akodon_dispersal <-  Herodotools::calc_dispersal_from(W = akodon_pa_tree,
                                                      tree = akodon_newick,
                                                      ancestral.area = node_area,
                                                      biogeo = biogeo_area)

#Load data
site_xy <- akodon_sites |>
  dplyr::select(LONG, LAT)

# coastline for maps
coastline <- rnaturalearth::ne_coastline(returnclass = "sf")

# defining map limits
map_limits <- list(
  x = c(-95, -30),
  y = c(-55, 12)
)

# just cropping coastline
coastline_crop <- sf::st_crop(coastline,
                              xmin = map_limits$x[1],
                              xmax = map_limits$x[2],
                              ymin = map_limits$y[1],
                              ymax = map_limits$y[2])


#Multiple maps
plot_dispersal_from(dispersal_from = akodon_dispersal,
                    coords = site_xy,
                    shapefile = coastline_crop,
                    area_cols = c(1, 3, 5),
                    facet = TRUE)

#All maps
plot_dispersal_from(akodon_dispersal,
                    coords = site_xy,
                    shapefile = coastline_crop,
                    facet = TRUE)
#' }

}
\author{
Maria Gabriela Junqueira and Gabriel Nakamura
}
