% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/function_spp_nodes.R
\name{spp_nodes}
\alias{spp_nodes}
\title{Species and their respective ancestral nodes - internal}
\usage{
spp_nodes(tree)
}
\arguments{
\item{tree}{phylogenetic tree}
}
\value{

}
\description{
Species and their respective ancestral nodes - internal
}
